Dear god this turned out complicated, but it works cross-browser!
