--- app/views/layouts/_flash_messages.html.haml
@@ -1,4 +1,0 @@
-- if notice
-  .alert.alert-success= notice
-- if alert
-  .alert.alert-danger= alert
