--- app/views/layouts/_navigation_bar.html.haml
@@ -1,23 +1,0 @@
-%nav.navbar{class: navbar_classes, role: "navigation"}
-  .container
-    %button.navbar-toggler.navbar-toggler-right{type: "button", "data-toggle" => "collapse", "data-target" => "#navbarSupportedContent", "aria-controls" => "navbarSupportedContent", "aria-expanded" => "false", "aria-label" => "Toggle navigation"}
-      %span.navbar-toggler-icon
-    = root_navigation_menus
-
-- unless browser.bot?
-  %noscript#no-javascript-warning
-    .container
-      .alert.alert-danger
-        %h2.mb-4 JavaScript disabled
-
-        .d-flex.align-items-center
-          .h1.m-0.mr-4
-            %i.fa.fa-exclamation-triangle
-          .flex-grow-1
-            %p
-              Your web browser has JavaScript disabled.  This site is written mostly in JavaScript,
-              and will not work without it.  Please enable JavaScript in your browser's settings (or
-              disable your JavaScript-blocking browser extension for this site).
-
-        .text-right
-          %a.btn.btn-primary{href: '.'} Reload page
